fix up tests make generateProgram its own thing
