+++ packages/core/src/shader/ShaderSystem.ts
@@ -1,8 +1,7 @@
 import { GLProgram } from './GLProgram';
-import { generateUniformsSync, unsafeEvalSupported, defaultValue, logProgramError, compileShader } from './utils';
+import { generateUniformsSync, unsafeEvalSupported } from './utils';
 
 import type { ISystem } from '../ISystem';
-import type { IGLUniformData } from './GLProgram';
 import type { Renderer } from '../Renderer';
 import type { IRenderingContext } from '../IRenderingContext';
 import type { Shader } from './Shader';
@@ -12,8 +11,7 @@ import type { Dict } from '@pixi/utils';
 import type { UniformsSyncCallback } from './utils';
 import { generateUniformBufferSync } from './utils/generateUniformBufferSync';
 
-import { getAttributeData } from './utils/getAttributeData';
-import { getUniformData } from './utils/getUniformData';
+import { generateProgram } from './utils/generateProgram';
 
 let UID = 0;
 // default sync data so we don't create a new one each time!
@@ -101,7 +99,7 @@ export class ShaderSystem implements ISystem
         shader.uniforms.globals = this.renderer.globalUniforms;
 
         const program = shader.program;
-        const glProgram = program.glPrograms[this.renderer.CONTEXT_UID] || this.generateShader(shader);
+        const glProgram = program.glPrograms[this.renderer.CONTEXT_UID] || this.generateProgram(shader);
 
         this.shader = shader;
 
@@ -308,59 +306,12 @@ export class ShaderSystem implements ISystem
      * @param {PIXI.Shader} shader - the shader that the glProgram will be based on.
      * @return {PIXI.GLProgram} A shiny new glProgram!
      */
-    generateShader(shader: Shader): GLProgram
+    generateProgram(shader: Shader): GLProgram
     {
         const gl = this.gl;
-
         const program = shader.program;
 
-        const glVertShader = compileShader(gl, gl.VERTEX_SHADER, program.vertexSrc);
-        const glFragShader = compileShader(gl, gl.FRAGMENT_SHADER, program.fragmentSrc);
-
-        const webGLProgram = gl.createProgram();
-
-        gl.attachShader(webGLProgram, glVertShader);
-        gl.attachShader(webGLProgram, glFragShader);
-
-        gl.linkProgram(webGLProgram);
-
-        if (!gl.getProgramParameter(webGLProgram, gl.LINK_STATUS))
-        {
-            logProgramError(gl, webGLProgram, glVertShader, glFragShader);
-        }
-
-        program.attributeData = getAttributeData(webGLProgram, gl);
-        program.uniformData = getUniformData(webGLProgram, gl);
-
-        const keys = Object.keys(program.attributeData);
-
-        keys.sort((a, b) => (a > b) ? 1 : -1); // eslint-disable-line no-confusing-arrow
-
-        for (let i = 0; i < keys.length; i++)
-        {
-            program.attributeData[keys[i]].location = i;
-
-            gl.bindAttribLocation(webGLProgram, i, keys[i]);
-        }
-
-        gl.linkProgram(webGLProgram);
-
-        gl.deleteShader(glVertShader);
-        gl.deleteShader(glFragShader);
-
-        const uniformData: {[key: string]: IGLUniformData} = {};
-
-        for (const i in program.uniformData)
-        {
-            const data = program.uniformData[i];
-
-            uniformData[i] = {
-                location: gl.getUniformLocation(webGLProgram, i),
-                value: defaultValue(data.type, data.size),
-            };
-        }
-
-        const glProgram = new GLProgram(webGLProgram, uniformData);
+        const glProgram = generateProgram(gl, program);
 
         program.glPrograms[this.renderer.CONTEXT_UID] = glProgram;
